refactor: 스토리상세 페이지의 params 네이밍을 id -> storyId로 변경
