+++ src/app/library/detail/[storyId]/_components/PaginatedContentViewer.tsx
